flatpak: Update for Tracker Miners 3
Tracker 3 is provided in the Flatpak SDK, see
https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/630.

We still build tracker-miners inside the Flatpak bundle, so that the
org.freedesktop.Tracker3.Miner.Files settings schema is available, and
so that the tests that depend on Tracker can pass as part of the CI
build.

Access to the host's miners is controlled by the new
org.freedesktop.Tracker.portal process provided in Tracker 3.